requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
psycopg2-binary>=2.9.0
opencv-python>=4.5.0
numpy>=1.19.0
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
easyocr>=1.6.0
boto3>=1.26.0
python-dotenv>=1.0.0
